requirements.txt
imapclient
pyzmail36
pushbullet.py
python-dotenv
